requirements.txt
numpy
pandas
matplotlib
praw
nltk
contractions
emoji
spacy
transformers
tensorflow
torch 
torchvision
boto3
python-dotenv
scikit-learn
seaborn
wordcloud
streamlit
plotly

-e .